+++ HISTORY.rst
@@ -4,9 +4,18 @@
 History
 -------
 
-4.9.0
+5.0.0
 ++++++++++++++++++
 
+* BREAKING: The ``raw`` attribute on the model classes has been replaced
+  with a ``to_dict()`` method. This can be used to get a representation of
+  the object that is suitable for serialization.
+* BREAKING: The ``ip_address`` property on the model classes now always returns
+  a ``ipaddress.IPv4Address`` or ``ipaddress.IPv6Address``.
+* BREAKING: The model and record classes now require all arguments other than
+  ``locales`` and ``ip_address`` to be keyword arguments.
+* BREAKING: ``geoip2.mixins`` has been made internal. This normally would not
+  have been used by external code.
 * IMPORTANT: Python 3.9 or greater is required. If you are using an older
   version, please use an earlier release.
 * ``metro_code`` on ``geoip2.record.Location`` has been deprecated. The